+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased changes
 機能追加
  * `g.Surface` でscale変更ができる設定を追加
- * `g.Surface` の省略可能なboolean型引数 `isDynamic` を booleanもしくはnumber型の `optionalFlag` に変更
+ * `g.Surface` の省略可能なboolean型引数 `isDynamic` を booleanもしくはnumber型の `statusOption` に変更
 
 ## 2.1.1
 不具合修正